+++ PayPalNativeCheckout/src/main/java/com/braintreepayments/api/PayPalNativeCheckoutAccountNonce.java
@@ -39,11 +39,11 @@ public class PayPalNativeCheckoutAccountNonce extends PaymentMethodNonce {
     private final String clientMetadataId;
     private final PostalAddress billingAddress;
     private final PostalAddress shippingAddress;
-    private final String firstName;
-    private final String lastName;
+    private String firstName;
+    private String lastName;
     private final String phone;
-    private final String email;
-    private final String payerId;
+    private String email;
+    private String payerId;
     private final PayPalNativeCheckoutCreditFinancing creditFinancing;
     private final String authenticateUrl;
 
@@ -196,6 +196,22 @@ public class PayPalNativeCheckoutAccountNonce extends PaymentMethodNonce {
     @NonNull
     public String getPayerId() {
         return payerId;
+    }
+
+    /**
+     * Sets the payer info if on the nonce object. This is used when the payer information is
+     * not returned from the json
+     *
+     * @param payerId - The payer id associated with the user
+     * @param firstName - The payer's first name
+     * @param lastName - The payer's last name
+     * @param email - The payer's email
+     */
+    public void setPayerInfo(String payerId, String firstName, String lastName, String email) {
+        this.payerId = payerId;
+        this.firstName = firstName;
+        this.lastName = lastName;
+        this.email = email;
     }
 
     /**