+++ PayPalNativeCheckout/src/main/java/com/braintreepayments/api/PayPalNativeCheckoutInternalClient.java
@@ -93,7 +93,6 @@ class PayPalNativeCheckoutInternalClient {
                 try {
                     PayPalNativeCheckoutAccountNonce payPalAccountNonce = PayPalNativeCheckoutAccountNonce.fromJSON(tokenizationResponse);
                     callback.onResult(payPalAccountNonce, null);
-
                 } catch (JSONException e) {
                     callback.onResult(null, e);
                 }