+++ CHANGELOG.md
@@ -38,6 +38,9 @@
     * `taxAmount`
     * `shippingAmount`
     * `lineItems`
+* PayPalNativeCheckout
+  * Bump native-checkout version to release `1.0.0`
+  * Fix an issue where the return from web fallback was not returning the correct information
 
 ## 4.31.0